requirements.txt
discord.py==0.16.6
aiohttp
websockets
chardet